+++ lib/ui/page/call/component/desktop.dart
@@ -30,7 +30,9 @@ import '../widget/call_cover.dart';
 import '../widget/conditional_backdrop.dart';
 import '../widget/dock.dart';
 import '../widget/hint.dart';
-import '../widget/participant.dart';
+import '../widget/participant/participant.dart';
+import '../widget/participant/participant_decorator.dart';
+import '../widget/participant/participant_overlay.dart';
 import '../widget/reorderable_fit.dart';
 import '../widget/scaler.dart';
 import '../widget/tooltip_button.dart';
