+++ lib/ui/page/call/component/mobile.dart
@@ -31,7 +31,9 @@ import '../widget/conditional_backdrop.dart';
 import '../widget/floating_fit/view.dart';
 import '../widget/hint.dart';
 import '../widget/minimizable_view.dart';
-import '../widget/participant.dart';
+import '../widget/participant/participant.dart';
+import '../widget/participant/participant_decorator.dart';
+import '../widget/participant/participant_overlay.dart';
 import '../widget/swappable_fit.dart';
 import '../widget/video_view.dart';
 import '/domain/model/avatar.dart';
